#127 - Refact: post page to use components created instead of embedding all info on same file
