+++ frontend/src/Pages/ShowPost/components/PostInfo/Description/descriptionSection.jsx
@@ -1,0 +1,22 @@
+import { Box, Paper } from '@mui/material';
+import { Descricao, Subtitulo } from '../../../style';
+
+const DescriptionSection = ({ text }) => {
+  return (
+    <Box
+      className="description"
+      sx={{
+        display: 'grid',
+        gap: '.5rem',
+        padding: '1rem',
+      }}
+    >
+      <Subtitulo>DESCRIÇÃO:</Subtitulo>
+      <Paper elevation={0}>
+        <Descricao>{text}</Descricao>
+      </Paper>
+    </Box>
+  );
+};
+
+export default DescriptionSection;
